Basic SVD implementation working (in SVD++fast.cpp). Ran with 50 features, K = 0.02, Lrate = 0.001, 157 epochs, got Quiz RMSE of 0.90706. Need to refactor code.

+++ globalMovieAvg.py
@@ -6,7 +6,7 @@ totalRating = 0.0
 
 for line in read_file:
 	vals = line.split()
-	rating = int(vals[3])
+	rating = float(vals[3])
 
 	if rating > 0:
 		numRatings += 1
